+++ Assets/Scripts/Tester.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using FMOD.Studio;
 using FMODUnity;
 using UnityEngine;
 
@@ -7,29 +8,32 @@ namespace ThanhDV.AudioManager.FMOD
 {
     public class Tester : MonoBehaviour
     {
-        [SerializeField] private EventReference bgmEventReference;
-        [SerializeField] private EventReference footEventReference;
+        [SerializeField] private EventReference _bgmEventReference;
+        [SerializeField] private EventReference _footEventReference;
+        [SerializeField] private EventReference _loopReference;
 
         void Start()
         {
-            AudioManager.Instance.PlayBGM("event:/Music/Level 02");
+            // AudioManager.Instance.PlayBGM("event:/Music/Level 02");
         }
 
         void Update()
         {
             if (Input.GetKeyDown(KeyCode.Alpha1))
             {
-                AudioManager.Instance.PlayOneShot("event:/Weapons/Pistol");
+                AudioManager.Instance.PlayLoop("Loop", _loopReference);
             }
 
             if (Input.GetKeyDown(KeyCode.Alpha2))
             {
-                AudioManager.Instance.PlayBGM(bgmEventReference);
+                AudioManager.Instance.TryGetEventInstance("Loop", out EventInstance instance);
+                instance.setPitch(2f);
             }
 
             if (Input.GetKeyDown(KeyCode.Alpha3))
             {
-                AudioManager.Instance.PlayBGM("event:/Music/Radio Station", 1f, 2f);
+                AudioManager.Instance.TryGetEventInstance("Loop", out EventInstance instance);
+                instance.setPitch(0.5f);
             }
 
             if (Input.GetKeyDown(KeyCode.Alpha4))
@@ -39,7 +43,7 @@ namespace ThanhDV.AudioManager.FMOD
 
             if (Input.GetKeyDown(KeyCode.Alpha5))
             {
-                AudioManager.Instance.PlayLoop("X", footEventReference, gameObject);
+                AudioManager.Instance.PlayLoop("X", _footEventReference, gameObject);
             }
 
             if (Input.GetKeyDown(KeyCode.Alpha6))
@@ -52,23 +56,23 @@ namespace ThanhDV.AudioManager.FMOD
                 Destroy(AudioManager.Instance.gameObject);
             }
 
-            if (Input.GetKeyDown(KeyCode.Alpha8))
-            {
-                float volume = AudioManager.Instance.GetVolume(AudioType.MASTER);
-                AudioManager.Instance.SetVolume(AudioType.MASTER, volume <= 0 ? 1f : 0f);
-            }
+            // if (Input.GetKeyDown(KeyCode.Alpha8))
+            // {
+            //     float volume = AudioManager.Instance.GetVolume(AudioType.MASTER);
+            //     AudioManager.Instance.SetVolume(AudioType.MASTER, volume <= 0 ? 1f : 0f);
+            // }
 
-            if (Input.GetKeyDown(KeyCode.Alpha9))
-            {
-                float volume = AudioManager.Instance.GetVolume(AudioType.BGM);
-                AudioManager.Instance.SetVolume(AudioType.BGM, volume <= 0 ? 1f : 0f);
-            }
+            // if (Input.GetKeyDown(KeyCode.Alpha9))
+            // {
+            //     float volume = AudioManager.Instance.GetVolume(AudioType.BGM);
+            //     AudioManager.Instance.SetVolume(AudioType.BGM, volume <= 0 ? 1f : 0f);
+            // }
 
-            if (Input.GetKeyDown(KeyCode.Alpha0))
-            {
-                float volume = AudioManager.Instance.GetVolume(AudioType.SFX);
-                AudioManager.Instance.SetVolume(AudioType.SFX, volume <= 0 ? 1f : 0f);
-            }
+            // if (Input.GetKeyDown(KeyCode.Alpha0))
+            // {
+            //     float volume = AudioManager.Instance.GetVolume(AudioType.SFX);
+            //     AudioManager.Instance.SetVolume(AudioType.SFX, volume <= 0 ? 1f : 0f);
+            // }
         }
     }
 }